Finished with BottomSheet Composable that is displayed when adding List, its gives the user 3 options, Add all items as unchecked, Add all items as is, and Add only unchecked items. The database call have not been implemented in ViewModel yet.
