+++ app/src/main/java/com/heyzeusv/yourlists/add/AddComposables.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.heightIn
@@ -16,15 +15,12 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExposedDropdownMenuBox
-import androidx.compose.material3.ExposedDropdownMenuDefaults
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Tab
 import androidx.compose.material3.TabRow
@@ -39,9 +35,7 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -98,6 +92,7 @@ fun AddScreen(
         addToListOnClick = addVM::addItem,
         deleteDefaultItemOnClick = addVM::deleteDefaultItem,
         itemLists = itemLists,
+        addListButtonOnClick = addVM::addListWithOption,
     )
 }
 
@@ -112,6 +107,7 @@ fun AddScreen(
     addToListOnClick: (DefaultItem) -> Unit,
     deleteDefaultItemOnClick: (DefaultItem) -> Unit,
     itemLists: List<ItemListWithItems>,
+    addListButtonOnClick: (ItemListWithItems, AddListOptions) -> Unit,
 ) {
     val pagerState = rememberPagerState(pageCount = { 2 })
     val coroutineScope = rememberCoroutineScope()
@@ -163,24 +159,37 @@ fun AddScreen(
         isVisible = showBottomSheet,
         updateIsVisible = { showBottomSheet = it },
     ) {
-        EditItemBottomSheetContent(
-            closeBottomSheet = {
-                updateDefaultItemQuery("")
-                showBottomSheet = false
-            },
-            selectedItem = selectedDefaultItem,
-            categories = categories,
-            primaryLabel = if (selectedDefaultItem.itemId == 0L) {
-                sRes(R.string.asbs_save_add)
-            } else {
-                sRes(R.string.asbs_update_add)
-            },
-            primaryOnClick = { saveAndAddOnClick(it as DefaultItem) },
-            secondaryLabel = sRes(R.string.asbs_add),
-            secondaryOnClick = { addToListOnClick(it as DefaultItem) },
-            deleteLabel = sRes(R.string.asbs_delete),
-            deleteOnClick = { deleteDefaultItemOnClick(it as DefaultItem) },
-        )
+        when (pagerState.currentPage) {
+            0 -> {
+                EditItemBottomSheetContent(
+                    closeBottomSheet = {
+                        updateDefaultItemQuery("")
+                        showBottomSheet = false
+                    },
+                    selectedItem = selectedDefaultItem,
+                    categories = categories,
+                    primaryLabel = if (selectedDefaultItem.itemId == 0L) {
+                        sRes(R.string.asbs_save_add)
+                    } else {
+                        sRes(R.string.asbs_update_add)
+                    },
+                    primaryOnClick = { saveAndAddOnClick(it as DefaultItem) },
+                    secondaryLabel = sRes(R.string.asbs_add),
+                    secondaryOnClick = { addToListOnClick(it as DefaultItem) },
+                    deleteLabel = sRes(R.string.asbs_delete),
+                    deleteOnClick = { deleteDefaultItemOnClick(it as DefaultItem) },
+                )
+            }
+            1 -> {
+                AddListBottomSheetContent(
+                    itemList = selectedItemList,
+                    buttonOnClick = { itemList, option ->
+                        addListButtonOnClick(itemList, option)
+                        showBottomSheet = false
+                    }
+                )
+            }
+        }
     }
 }
 
@@ -309,70 +318,45 @@ fun AddListPage(
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun FilteredDropDownMenu(
-    value: TextFieldValue,
-    onValueChanged: (TextFieldValue) -> Unit,
-    label: String,
-    options: List<String>,
-    optionOnClick: (String) -> Unit,
-    maxLength: Int,
-    modifier: Modifier = Modifier,
-    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
-    keyboardActions: KeyboardActions = KeyboardActions.Default,
+fun AddListBottomSheetContent(
+    itemList: ItemListWithItems,
+    buttonOnClick: (ItemListWithItems, AddListOptions) -> Unit,
 ) {
-    var expanded by remember { mutableStateOf(false) }
-
-    ExposedDropdownMenuBox(
-        expanded = expanded,
-        onExpandedChange = { expanded = !expanded },
-        modifier = modifier,
+    Column(
+        modifier = Modifier
+            .padding(all = dRes(R.dimen.bs_padding_all))
+            .fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(dRes(R.dimen.bs_horizontal_spacedBy)),
     ) {
-        TextField(
-            value = value,
-            onValueChange = { if (it.text.length <= maxLength) onValueChanged(it) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .onFocusChanged { expanded = it.isFocused }
-                .menuAnchor(),
-            label = { Text(label) },
-            trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) },
-            supportingText = {
-                Row {
-                    Text(
-                        text = "${value.text.length}/$maxLength",
-                        modifier = Modifier.fillMaxWidth(),
-                        textAlign = TextAlign.End,
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                }
-            },
-            keyboardOptions = keyboardOptions,
-            keyboardActions = keyboardActions,
-            singleLine = true,
+        Text(
+            text = itemList.itemList.name,
+            style = MaterialTheme.typography.headlineMedium
         )
-        val filteredOptions = options.filter { it.contains(value.text, ignoreCase = true) }
-        if (filteredOptions.isNotEmpty()) {
-            ExposedDropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { }
-            ) {
-                filteredOptions.forEach { option ->
-                    DropdownMenuItem(
-                        text = {
-                            Text(
-                                text = option,
-                                maxLines = 1,
-                            )
-                        },
-                        onClick = {
-                            optionOnClick(option)
-                            expanded = false
-                        },
-                    )
-                }
-            }
+        Button(
+            onClick = { buttonOnClick(itemList, AddListOptions.ALL_AS_UNCHECKED) },
+            modifier = Modifier.fillMaxWidth(),
+            shape = MaterialTheme.shapes.extraSmall,
+        ) {
+            Text(text = sRes(R.string.asbs_add_all_unchecked).uppercase())
+        }
+        Button(
+            onClick = { buttonOnClick(itemList, AddListOptions.ALL_AS_IS) },
+            modifier = Modifier.fillMaxWidth(),
+            shape = MaterialTheme.shapes.extraSmall,
+            colors = ButtonDefaults.filledTonalButtonColors(
+                containerColor = MaterialTheme.colorScheme.primaryContainer,
+                contentColor = MaterialTheme.colorScheme.onPrimaryContainer
+            )
+        ) {
+            Text(text = sRes(R.string.asbs_add_all_as_is).uppercase())
+        }
+        OutlinedButton(
+            onClick = { buttonOnClick(itemList, AddListOptions.ONLY_UNCHECKED) },
+            modifier = Modifier.fillMaxWidth(),
+            shape = MaterialTheme.shapes.extraSmall,
+        ) {
+            Text(text = sRes(R.string.asbs_add_only_uncheck).uppercase())
         }
     }
 }
@@ -387,10 +371,11 @@ private fun AddScreenPreview() {
                 updateDefaultItemQuery = { },
                 defaultItems = defaultItemList,
                 categories = emptyList(),
-                itemLists = emptyList(),
                 saveAndAddOnClick = { },
                 addToListOnClick = { },
                 deleteDefaultItemOnClick = { },
+                itemLists = emptyList(),
+                addListButtonOnClick = { _, _ -> },
             )
         }
     }
@@ -406,10 +391,11 @@ private fun AddScreenBlankQueryPreview() {
                 updateDefaultItemQuery = { },
                 defaultItems = defaultItemList,
                 categories = emptyList(),
-                itemLists = emptyList(),
                 saveAndAddOnClick = { },
                 addToListOnClick = { },
                 deleteDefaultItemOnClick = { },
+                itemLists = emptyList(),
+                addListButtonOnClick = { _, _ -> },
             )
         }
     }
@@ -446,17 +432,15 @@ private fun AddListPagePreview() {
 
 @Preview
 @Composable
-private fun FilteredDropDownMenuPreview() {
+private fun AddListBottomSheetPreview() {
     PreviewUtil.run {
         Preview {
-            FilteredDropDownMenu(
-                value = TextFieldValue("Preview"),
-                onValueChanged = { },
-                label = "Preview Label",
-                options = listOf("Preview1", "Preview2", "Preview3"),
-                optionOnClick = { },
-                maxLength = 999,
-            )
+            Surface {
+                AddListBottomSheetContent(
+                    itemList = halfCheckedItemList,
+                    buttonOnClick = { _, _ -> },
+                )
+            }
         }
     }
 }
