+++ app/src/main/java/com/heyzeusv/yourlists/add/AddListOptions.kt
@@ -1,0 +1,7 @@
+package com.heyzeusv.yourlists.add
+
+enum class AddListOptions {
+    ALL_AS_UNCHECKED,
+    ALL_AS_IS,
+    ONLY_UNCHECKED,
+}
